refactor: complete server.go split (types/render/middleware/session/helpers/settings_health/templates_compile) + feeds/srs split
