+++ internal/feeds/srs_json.go
@@ -1,0 +1,5 @@
+package feeds
+
+// srs_json.go — JSON-based SRS adapter parsing.
+// Currently no JSON-format SRS data is supported; all SRS parsing uses the
+// sing-box binary format via ParseSRS in srs.go.
