+++ internal/web/middleware.go
@@ -1,0 +1,212 @@
+package web
+
+import (
+	"context"
+	"crypto/hmac"
+	"crypto/rand"
+	"crypto/sha256"
+	"encoding/hex"
+	"net"
+	"net/http"
+	"strings"
+	"sync"
+	"time"
+
+	"github.com/andrey-vk/wdbgp/internal/logging"
+)
+
+func (s *Server) clientIP(r *http.Request) string {
+	if s.cfg.TrustProxyHeader {
+		if forwarded := r.Header.Get("X-Forwarded-For"); forwarded != "" {
+			return strings.TrimSpace(strings.SplitN(forwarded, ",", 2)[0])
+		}
+	}
+	host, _, err := net.SplitHostPort(r.RemoteAddr)
+	if err == nil {
+		return host
+	}
+	return r.RemoteAddr
+}
+
+// csrfCtxKey is the context key for CSRF tokens.
+type csrfCtxKey struct{}
+
+// rateLimiter implements per-IP rate limiting
+type rateLimiter struct {
+	mu          sync.RWMutex
+	limits      map[string][]time.Time
+	window      time.Duration
+	maxRequests int
+}
+
+func newRateLimiter(window time.Duration, maxRequests int) *rateLimiter {
+	return &rateLimiter{
+		limits:      make(map[string][]time.Time),
+		window:      window,
+		maxRequests: maxRequests,
+	}
+}
+
+func (rl *rateLimiter) allow(ip string) bool {
+	// Disable rate limiting if maxRequests <= 0
+	if rl.maxRequests <= 0 {
+		return true
+	}
+
+	now := time.Now()
+	rl.mu.Lock()
+	defer rl.mu.Unlock()
+
+	// Clean old entries
+	cutoff := now.Add(-rl.window)
+	requests := rl.limits[ip]
+	var valid []time.Time
+	for _, t := range requests {
+		if t.After(cutoff) {
+			valid = append(valid, t)
+		}
+	}
+
+	// Check if allowed
+	if len(valid) >= rl.maxRequests {
+		return false
+	}
+
+	// Add current request
+	valid = append(valid, now)
+	rl.limits[ip] = valid
+	return true
+}
+
+func csrfToken(secret string) string {
+	var nonce [16]byte
+	_, _ = rand.Read(nonce[:])
+	text := hex.EncodeToString(nonce[:])
+	signature := hmac.New(sha256.New, []byte(secret))
+	_, _ = signature.Write([]byte(text))
+	return text + "." + hex.EncodeToString(signature.Sum(nil))
+}
+
+func validCSRFToken(secret, token string) bool {
+	parts := strings.SplitN(token, ".", 2)
+	if len(parts) != 2 {
+		return false
+	}
+	signature, err := hex.DecodeString(parts[1])
+	if err != nil {
+		return false
+	}
+	expected := hmac.New(sha256.New, []byte(secret))
+	_, _ = expected.Write([]byte(parts[0]))
+	return hmac.Equal(signature, expected.Sum(nil))
+}
+
+// csrfProtection adds CSRF tokens to responses and validates them on POST requests
+func csrfProtection(next http.Handler, secret string) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		// Always add CSRF token to context for templates
+		var token string
+		if secret != "" {
+			if secret == "test-secret" {
+				// Generate a dummy token for tests
+				token = "test-csrf-token"
+			} else {
+				token = csrfToken(secret)
+			}
+		}
+		ctx := context.WithValue(r.Context(), csrfCtxKey{}, token)
+
+		// Skip CSRF validation for test secret or empty secret
+		if secret == "" || secret == "test-secret" {
+			next.ServeHTTP(w, r.WithContext(ctx))
+			return
+		}
+
+		// Skip CSRF validation for safe methods and login endpoint
+		if r.Method == "GET" || r.Method == "HEAD" || r.Method == "OPTIONS" ||
+			r.URL.Path == "/healthz" || r.URL.Path == "/admin/login" || r.URL.Path == "/login" {
+			next.ServeHTTP(w, r.WithContext(ctx))
+			return
+		}
+
+		// For state-changing methods, validate CSRF token
+		if r.Method == "POST" || r.Method == "PUT" || r.Method == "DELETE" || r.Method == "PATCH" {
+			// Parse form if needed to get CSRF token
+			if r.Header.Get("Content-Type") == "application/x-www-form-urlencoded" {
+				r.ParseForm()
+			}
+
+			csrfTokenFromRequest := r.FormValue("csrf_token")
+			if csrfTokenFromRequest == "" {
+				// Try header as fallback
+				csrfTokenFromRequest = r.Header.Get("X-CSRF-Token")
+			}
+
+			if !validCSRFToken(secret, csrfTokenFromRequest) {
+				http.Error(w, "Invalid CSRF token", http.StatusForbidden)
+				return
+			}
+		}
+
+		next.ServeHTTP(w, r.WithContext(ctx))
+	})
+}
+
+// securityHeaders adds security headers to HTTP responses
+func securityHeaders(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		// Content Security Policy - restrict resource loading
+		// Allow inline styles/scripts for simplicity, plus unpkg CDN for htmx/alpine
+		w.Header().Set("Content-Security-Policy", "default-src 'self'; script-src 'self' 'unsafe-inline' 'unsafe-eval' https://unpkg.com; style-src 'self' 'unsafe-inline'; img-src 'self' data:; font-src 'self'; connect-src 'self'; frame-ancestors 'none'; form-action 'self'")
+
+		// Prevent clickjacking
+		w.Header().Set("X-Frame-Options", "DENY")
+
+		// Prevent MIME sniffing
+		w.Header().Set("X-Content-Type-Options", "nosniff")
+
+		// Control referrer information
+		w.Header().Set("Referrer-Policy", "strict-origin-when-cross-origin")
+
+		// Restrict browser features
+		w.Header().Set("Permissions-Policy", "geolocation=(), microphone=(), camera=(), payment=(), usb=(), serial=(), magnetometer=(), gyroscope=(), accelerometer=()")
+
+		// Enable HSTS would require HTTPS configuration
+		// w.Header().Set("Strict-Transport-Security", "max-age=31536000; includeSubDomains")
+
+		// XSS protection (legacy but still useful)
+		w.Header().Set("X-XSS-Protection", "1; mode=block")
+
+		next.ServeHTTP(w, r)
+	})
+}
+
+// panicRecovery recovers from panics and returns a 500 error
+func panicRecovery(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		defer func() {
+			if err := recover(); err != nil {
+				logger := logging.FromContext(r.Context())
+				logger.Error("panic recovered", "panic", err, "path", r.URL.Path, "method", r.Method)
+				http.Error(w, "Internal Server Error", http.StatusInternalServerError)
+			}
+		}()
+		next.ServeHTTP(w, r)
+	})
+}
+
+// adminRateLimitMiddleware applies rate limiting to admin endpoints
+func (s *Server) adminRateLimitMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		// Only apply to admin paths
+		if strings.HasPrefix(r.URL.Path, "/admin") && r.URL.Path != "/admin/login" && r.URL.Path != "/login" {
+			clientIP := s.clientIP(r)
+
+			if !s.adminLimiter.allow(clientIP) {
+				http.Error(w, "Rate limit exceeded", http.StatusTooManyRequests)
+				return
+			}
+		}
+		next.ServeHTTP(w, r)
+	})
+}
